+++ tools/lgtdoc/xml/lgt2txt.js
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////
 //
 //   XML documenting files to plain text conversion script 
-//   Last updated on July 5, 2016
+//   Last updated on July 30, 2016
 //
 //   This file is part of Logtalk <https://logtalk.org/>  
 //   Copyright 1998-2021 Paulo Moura <pmoura@logtalk.org>
@@ -35,6 +35,7 @@ var processor = "msxsl";
 // var processor = "xsltproc";
 // var processor = "xalan";
 // var processor = "sabcmd";
+// var processor = "saxon";
 
 if (WScript.Arguments.Unnamed.Length > 0) {
 	usage_help();
@@ -88,7 +89,7 @@ if (d_arg != "" && !FSObject.FolderExists(d_arg)) {
 } else if (d_arg != "")
 	directory = d_arg;
 
-if (p_arg != "" && p_arg != "msxsl" && p_arg != "xsltproc" && p_arg != "xalan" && p_arg != "sabcmd") {
+if (p_arg != "" && p_arg != "msxsl" && p_arg != "xsltproc" && p_arg != "xalan" && p_arg != "sabcmd" && p_arg != "saxon") {
 	WScript.Echo("Error! Unsupported XSLT processor:" + p_arg);
 	WScript.Echo("");
 	usage_help();
@@ -148,6 +149,9 @@ for (files.moveFirst(); !files.atEnd(); files.moveNext()) {
 			case "sabcmd" :
 				WshShell.Run("sabcmd \"" + xslt + "\" \"" + file + "\" \"" + txt_file + "\"", true);
 				break;
+			case "saxon" :
+				WshShell.Run("java net.sf.saxon.Transform -o:\"" + txt_file + "\" -s:\"" + file + "\" -xsl:\"" + xslt + "\"", true);
+				break;
 		}
 	}
 }
@@ -168,7 +172,7 @@ function usage_help() {
 	WScript.Echo("");
 	WScript.Echo("Optional arguments:");
 	WScript.Echo("  d - output directory for the generated files (default is " + directory + ")");
-	WScript.Echo("  p - XSLT processor (msxsl, xsltproc, xalan, or sabcmd; default is " + processor + ")");
+	WScript.Echo("  p - XSLT processor (msxsl, xsltproc, xalan, sabcmd, or saxon; default is " + processor + ")");
 	WScript.Echo("");
 	WScript.Quit(1);
 }